Created RoomCodeTest & Updated NetworkController.cs
Test for the RoomCode is executed via a BASH script in the Assets Directory.
The script is titled 'RoomCodeTest' and checks the RoomCode.txt file's contents to see if the RoomCode is appropriate.

The NetworkController.cs file was updated - specifically the Start() method - in order to output the generated Room Code into the file RoomCode.txt.